RMI Interface with Store

See merge request cpd/2122/t09/g01!11

+++ Dynamo/src/server/cluster/MembershipService.java
@@ -4,12 +4,16 @@ import common.Message;
 import common.Sender;
 import common.Utils;
 import server.Constants;
-
+import server.network.TCPListener;
 import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.util.*;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Future;
 
 public class MembershipService implements ClusterMembership {
     private final TreeMap<String, Node> nodeMap;
@@ -21,12 +25,14 @@ public class MembershipService implements ClusterMembership {
     private final String folderPath;
     private int membershipCounter = 0; // NEEDS TO BE STORED IN NON-VOLATILE MEMORY TO SURVIVE NODE CRASHES
     private static final int maxRetransmissions = 3;
-    private int retransmissionCounter = 0;
     private final HashSet<String> membershipReplyNodes;
     private final HashSet<String> repliedNodes;
 
+    private boolean isElected = false;
+    Future<?> electionPingThread = null;
+
     public MembershipService(String multicastIPAddr, int multicastIPPort, String nodeId, int tcpPort) {
-        nodeMap = new TreeMap<>();
+        this.nodeMap = new TreeMap<>();
         this.multicastIpAddr = multicastIPAddr;
         this.multicastIPPort = multicastIPPort;
         this.nodeId = nodeId;
@@ -37,30 +43,41 @@ public class MembershipService implements ClusterMembership {
         this.createNodeFolder();
     }
 
+    /**
+     * This methods throws a RuntimeException if it fails
+     */
     @Override
     public void join() {
         if (this.membershipCounter == 0 || !isClusterMember(this.membershipCounter)) {
             if (this.membershipCounter != 0) // Edge case for the first join
                 this.updateMembershipCounter(this.membershipCounter + 1);
-            this.addLog(this.nodeId, this.membershipCounter);
+
             this.multicastJoin();
-            // TODO: OPEN UDP / TCP LISTENER ??
+
+            // Send election request
+            ElectionService.sendRequest(this.nodeId, this.nodeMap);
         } else {
             throw new RuntimeException("Attempting to join the cluster while being already a member.");
         }
     }
 
+    /**
+     * This methods throws a RuntimeException if it fails
+     */
     @Override
     public void leave() {
-        // TODO Leave protocol
         if (isClusterMember(this.membershipCounter)) {
             this.removeNodeFromMap(this.nodeId);
             this.updateMembershipCounter(this.membershipCounter + 1);
-            this.addLog(this.nodeId, this.membershipCounter);
+            this.addLog(this.nodeId, this.membershipCounter, this.tcpPort);
             this.multicastLeave();
-            // TODO: CLOSE UDP / TCP LISTENER ??
-        } else {
-            throw new RuntimeException("Attempting to leave the cluster while not being a member.");
+
+            if (this.isElected) {
+                ElectionService.sendLeave(this.nodeMap, this.buildMembershipMsgBody());
+
+                this.isElected = false;
+                if (this.electionPingThread != null) this.electionPingThread.cancel(true);
+            }
         }
     }
 
@@ -94,9 +111,10 @@ public class MembershipService implements ClusterMembership {
 
         // Add this node information
         this.addNodeToMap(this.nodeId, this.tcpPort);
-        this.addLog(this.nodeId, this.membershipCounter);
+        this.addLog(this.nodeId, this.membershipCounter, this.tcpPort);
 
-        while (this.retransmissionCounter < maxRetransmissions) {
+        int retransmissionCounter = 0;
+        while (retransmissionCounter < maxRetransmissions) {
             int elapsedTime = 0;
             try {
                 Sender.sendMulticast(msg.toBytes(), this.multicastIpAddr, this.multicastIPPort);
@@ -104,7 +122,6 @@ public class MembershipService implements ClusterMembership {
                     Thread.sleep(Constants.multicastStepTime);
                     if (this.membershipReplyNodes.size() >= Constants.numMembershipMessages) {
                         this.membershipReplyNodes.clear();
-                        this.retransmissionCounter = 0;
                         System.out.println("New Node joined the distributed store");
                         return;
                     }
@@ -115,7 +132,7 @@ public class MembershipService implements ClusterMembership {
                 throw new RuntimeException(e);
             }
 
-            this.retransmissionCounter++;
+            retransmissionCounter++;
         }
 
         System.out.println("Prime Node joined the distributed store.");
@@ -152,6 +169,8 @@ public class MembershipService implements ClusterMembership {
     }
 
     /**
+     * Beware that when counter is 0 it can or not be a member
+     * @param counter
      * @return true if a node is a member of the nodeMap. False otherwise.
      */
     public static boolean isClusterMember(int counter) {
@@ -228,11 +247,12 @@ public class MembershipService implements ClusterMembership {
 
             synchronized (logPath.intern()) {
                 File memberLog = new File(logPath);
-
-                // Set initial log to be the current node
-                FileWriter writer = new FileWriter(memberLog, false);
-                writer.write(String.format("%s %d", this.nodeId, this.membershipCounter));
-                writer.close();
+                if (!memberLog.exists()) {
+                    // Set initial log to be the current node
+                    FileWriter writer = new FileWriter(memberLog, false);
+                    writer.write(String.format("%s %d", this.nodeId, this.membershipCounter));
+                    writer.close();
+                }
             }
         } catch (IOException e) {
             throw new RuntimeException(e);
@@ -243,11 +263,11 @@ public class MembershipService implements ClusterMembership {
      * Adds a new log to the beginning of the membership log removing existing logs
      * for that nodeId
      */
-    public void addLog(String newNodeId, int newMemberCounter) {
+    public void addLog(String newNodeId, int newMemberCounter, int newNodePort) {
         String logPath = this.folderPath + Constants.membershipLogFileName;
 
-        synchronized (logPath.intern()) {
-            try {
+        try {
+            synchronized (logPath.intern()) {
                 File file = new File(logPath);
                 boolean isDeprecatedLog = false;
 
@@ -280,19 +300,22 @@ public class MembershipService implements ClusterMembership {
 
                     Files.write(file.toPath(), filteredFile, StandardOpenOption.WRITE,
                             StandardOpenOption.TRUNCATE_EXISTING);
-
-                    // If the newMemberCounter is even, it is already added by the nodeMap received
-                    if (!isClusterMember(newMemberCounter)) {
-                        // Remove the node from the nodeMap
-                        this.removeNodeFromMap(newNodeId);
-                    }
-                }
-            } catch (FileNotFoundException e) {
-                System.out.println("An error occurred.");
-                e.printStackTrace();
-            } catch (IOException e) {
-                e.printStackTrace();
+                } 
             }
+
+            if (isClusterMember(newMemberCounter)) {
+                // if the nodeMap does not contain this node
+                if ((newNodePort != Constants.invalidPort) && !this.nodeMap.containsKey(Utils.generateKey(newNodeId)))
+                    this.nodeMap.put(Utils.generateKey(newNodeId), new Node(newNodeId, newNodePort));
+            } else {
+                // Remove the node from the nodeMap
+                this.removeNodeFromMap(newNodeId);
+            }
+        } catch (FileNotFoundException e) {
+            System.out.println("An error occurred.");
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
         }
     }
 
@@ -304,25 +327,12 @@ public class MembershipService implements ClusterMembership {
             byteOut.write(this.nodeId.getBytes(StandardCharsets.UTF_8));
             byteOut.write(Utils.newLine.getBytes(StandardCharsets.UTF_8));
 
-            String logPath = this.folderPath + Constants.membershipLogFileName;
-
-            synchronized (logPath.intern()) {
-                File file = new File(logPath);
-                Scanner myReader = new Scanner(file);
-                for (int i = 0; i < Constants.numLogEvents; i++) {
-                    if (!myReader.hasNextLine())
-                        break;
-                    String line = myReader.nextLine();
-                    byteOut.write(line.getBytes(StandardCharsets.UTF_8));
-                    byteOut.write(Utils.newLine.getBytes(StandardCharsets.UTF_8));
-                }
-                myReader.close();
-            }
+            byteOut.write(LogHandler.buildLogsBytes(this.folderPath, null));
 
             byteOut.write(Utils.newLine.getBytes(StandardCharsets.UTF_8));
 
             for (Node node : this.getNodeMap().values()) {
-                String entryLine = node.getId() + " " + node.getPort() + "\n";
+                String entryLine = node.getId() + " " + node.getPort() + Utils.newLine;
                 byteOut.write(entryLine.getBytes(StandardCharsets.UTF_8));
             }
         } catch (IOException e) {
@@ -344,7 +354,7 @@ public class MembershipService implements ClusterMembership {
 
         // Updates view of the cluster membership and adds the log
         this.addNodeToMap(nodeId, tcpPort);
-        this.addLog(nodeId, membershipCounter);
+        this.addLog(nodeId, membershipCounter, tcpPort);
 
         // Updated membership info TODO: CHECK IF THIS IS OK
         this.repliedNodes.clear();
@@ -362,6 +372,15 @@ public class MembershipService implements ClusterMembership {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    public void handleLeaveRequest(String nodeId,  int membershipCounter, int tcpPort) {
+        // Updates view of the cluster membership and adds the log
+        this.removeNodeFromMap(nodeId);
+        this.addLog(nodeId, membershipCounter, tcpPort);
+
+        // Updated membership info TODO: CHECK IF THIS IS OK
+        this.repliedNodes.clear();
     }
 
     public void handleMembershipResponse(Message message) {
@@ -394,13 +413,184 @@ public class MembershipService implements ClusterMembership {
             throw new RuntimeException(e);
         }
 
-        for (String newLog : membershipLogs) {
+        updateMembershipInfo(membershipLogs);
+
+        System.out.println("Received membership Logs: " + membershipLogs + "\nnodeMap: " + this.getNodeMap());
+    }
+
+    public int getMembershipCounter() {
+        return membershipCounter;
+    }
+
+    /**
+     *
+     * @return true if node crashed while being a member of the cluster
+     */
+    public boolean hasCrashed() {
+        if (!isClusterMember(this.membershipCounter)) return false;
+
+        // In case a cluster is composed by only 1 node joining and leaving
+        if (this.membershipCounter != 0) return true;
+
+        try {
+            Path path = Paths.get(this.folderPath + Constants.membershipLogFileName);
+            return Files.lines(path).count() > 1;
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public String getNodeId() {
+        return nodeId;
+    }
+
+    public void handleElectionRequest(Message message, ExecutorService executorService) {
+
+        InputStream is = new ByteArrayInputStream(message.getBody());
+        BufferedReader br = new BufferedReader(new InputStreamReader(is));
+
+        String line, newNodeId;
+        final HashMap<String, Integer> membershipLogs = new HashMap<>();
+        try {
+            newNodeId = br.readLine();
+            // Verify if newNodeId received is this node (meaning this node was elected)
+            if (newNodeId.equals(this.nodeId) && !this.isElected) {
+                 if (executorService != null) {
+                     this.electionPingThread = executorService.submit(new ElectionService(this.nodeId, this.folderPath, this.multicastIpAddr, this.multicastIPPort, this.nodeMap));
+                     this.isElected = true;
+                     System.out.println("THIS NODE WAS ELECTED");
+                 }
+                return;
+            }
+
+            System.out.printf("Received election request from %s\n", newNodeId);
+            while ((line = br.readLine()) != null) {
+                String[] logData = line.split(" ");
+                membershipLogs.put(logData[0], Integer.parseInt(logData[1]));
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        if (LogHandler.shouldPropagate(membershipLogs, newNodeId, this.folderPath, this.nodeId)) {
+            // Propagate the message to the next node?
+            System.out.println("Log is more recent! propagate to next node");
+
+            // Check if this node was the previous leader
+            if (this.isElected) {
+                this.isElected = false;
+                if (this.electionPingThread != null) this.electionPingThread.cancel(true);
+            }
+
+            // Send election request
+            ElectionService.propagateRequest(this.nodeId, this.nodeMap, message);
+        }
+    }
+
+    public void handleElectionPing(Message message) {
+        System.out.println("Received election ping.");
+
+        ByteArrayInputStream is = new ByteArrayInputStream(message.getBody());
+        BufferedReader br = new BufferedReader(new InputStreamReader(is));
+
+        String line, newNodeId;
+        final HashMap<String, ArrayList<Integer>> newMembershipLogs = new HashMap<>();
+        final HashMap<String, Integer> newParsedLogs = new HashMap<>();
+        try {
+            newNodeId = br.readLine();
+
+            while ((line = br.readLine()) != null) {
+                if (line.isEmpty())
+                    break;
+                String[] logData = line.split(" ");
+                newMembershipLogs.put(logData[0], new ArrayList<>(List.of( Integer.parseInt(logData[1]), Integer.parseInt(logData[2]) )));
+                newParsedLogs.put(logData[0], Integer.parseInt(logData[1]));
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        // Update current node logs
+        final HashMap<String, Integer> currMembershipLogs = LogHandler.buildLogsMap(this.folderPath, Integer.MAX_VALUE);
+        for (String iterNodeId : newMembershipLogs.keySet()) {
+            int currCounter = -1;
+            if (currMembershipLogs.containsKey(iterNodeId))
+                currCounter = currMembershipLogs.get(iterNodeId);
+
+            Integer newCounter = newParsedLogs.get(iterNodeId);
+            if (newCounter > currCounter) {
+                this.addLog(iterNodeId, newCounter, newMembershipLogs.get(iterNodeId).get(1));
+            }
+        }
+
+        if (LogHandler.shouldBeElected(newParsedLogs, this.folderPath)) {
+            // Propagate the message to the next node?
+            System.out.println("Node is more recent than the current leader. Starting an election request...");
+
+            // Send election request
+            ElectionService.sendRequest(this.nodeId, this.nodeMap);
+        }
+    }
+    public void handleElectionLeave(Message message) {
+        System.out.println("Received election leave.");
+
+        InputStream is = new ByteArrayInputStream(message.getBody());
+        BufferedReader br = new BufferedReader(new InputStreamReader(is));
+
+        String line;
+        final ArrayList<String> newMembershipLogs = new ArrayList<>();
+        try {
+            br.readLine();
+
+            while ((line = br.readLine()) != null) {
+                if (line.isEmpty())
+                    break; // Reached the end of membership log
+                newMembershipLogs.add(line);
+            }
+
+            while ((line = br.readLine()) != null) {
+                String[] data = line.split(" ");
+                String newNodeId = data[0];
+                int newNodePort = Integer.parseInt(data[1]);
+                this.addNodeToMap(newNodeId, newNodePort); // Check what happens when adding node that already exists
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        updateMembershipInfo(newMembershipLogs);
+
+        // Send election request to become the new leader
+        ElectionService.sendRequest(this.nodeId, this.nodeMap);
+    }
+
+    public void handleElectionTimeout() {
+        System.out.println("Election Ping timeout detected! Sending an election request...");
+        ElectionService.sendRequest(this.nodeId, this.nodeMap);
+    }
+
+    /**
+     * This method updates this node membershipInfo efficiently, according to the new received logs
+     * @param newMembershipLogs
+     */
+    private void updateMembershipInfo(ArrayList<String> newMembershipLogs) {
+        HashMap<String, Integer> currMembershipLogs = LogHandler.buildLogsMap(this.folderPath, Integer.MAX_VALUE);
+
+        for (String newLog : newMembershipLogs) {
             String[] logData = newLog.split(" ");
             String logId = logData[0];
             int logCounter = Integer.parseInt(logData[1]);
-            this.addLog(logId, logCounter);
-        }
 
-        System.out.println("Received membership Logs: " + membershipLogs + "\nnodeMap: " + this.getNodeMap());
+            int currCounter = -1;
+            if (currMembershipLogs.containsKey(logId))
+                currCounter = currMembershipLogs.get(logId);
+
+            if (logCounter > currCounter)
+                this.addLog(logId, logCounter, Constants.invalidPort);
+        }
+    }
+
+    public boolean getIsElected() {
+        return this.isElected;
     }
 }
